Move HashBox into separate file
