+++ Wire-iOS/Sources/Helpers/HashBox.swift
@@ -1,0 +1,39 @@
+//
+// Wire
+// Copyright (C) 2020 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+/// A wrapper type to provide hashable capabilities for abstract types.
+
+final class HashBox<Type: NSObjectProtocol>: Hashable {
+
+    let value: Type
+
+    init(value: Type) {
+        self.value = value
+    }
+
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(value.hash)
+    }
+
+    static func == (lhs: HashBox<Type>, rhs: HashBox<Type>) -> Bool {
+        return lhs.value.isEqual(rhs.value)
+    }
+
+}
